+++ Coursework/Front/front-end/src/pages/AccountPage.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import TopBar from './components/TopBar';
-import NavBar from './components/NavBar';
-import WorkerCard from './components/WorkerCard';
+import TopBar from '../components/TopBar';
+import NavBar from '../components/NavBar';
+import WorkerCard from '../components/WorkerCard';
 
 export default class AccountPage extends React.Component<any, any> {
   public render() {
