+++ Coursework/Front/front-end/src/pages/LearningPage.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import TopBar from './components/TopBar';
-import NavBar from './components/NavBar';
-import LearningList from './components/LearningList';
+import TopBar from '../components/TopBar';
+import NavBar from '../components/NavBar';
+import LearningList from '../components/LearningList';
 
 export default class LearningPage extends React.Component<any, any> {
   public render() {
